Big refactoring of types, and hopefully improvement

+++ Plato.Geometry.CSharp/Graphics/Material.cs
@@ -7,16 +7,16 @@ namespace Plato.Geometry.Graphics
     {
         public string Name;
         public Color Color;
-        public double Transmission;
-        public double Metallic;
-        public double Roughness;
-        public double Sheen;
-        public double SheenTint;
+        public float Transmission;
+        public float Metallic;
+        public float Roughness;
+        public float Sheen;
+        public float SheenTint;
         public bool UseSpecularReflectance = false;
-        public double SpecularReflectance = 0.4;
-        public double IndexOfRefraction;
-        public double SpecularTint;
-        public double SubsurfaceScattering;
+        public float SpecularReflectance = 0.4;
+        public float IndexOfRefraction;
+        public float SpecularTint;
+        public float SubsurfaceScattering;
         public Color SubsurfaceColor;
 
         public Material(Color color)
@@ -27,23 +27,23 @@ namespace Plato.Geometry.Graphics
             int green, 
             int blue, 
             int alpha,            
-            double transmission, 
-            double metallic, 
-            double roughness, 
-            double sheen, 
-            double sheenTint, 
+            float transmission, 
+            float metallic, 
+            float roughness, 
+            float sheen, 
+            float sheenTint, 
             bool useSpecularReflectance, 
-            double specularReflectance, 
-            double indexOfRefraction, 
-            double specularTint, 
-            double subsurfaceScattering, 
+            float specularReflectance, 
+            float indexOfRefraction, 
+            float specularTint, 
+            float subsurfaceScattering, 
             int subsurfaceRed,
             int subsurfaceGreen,
             int subsurfaceBlue,
             [System.Runtime.CompilerServices.CallerMemberName] string name = "")
         {
             Debug.Assert(alpha == 255, "Alpha channel is not currently supported");
-            Color = new Color(red / 255.0, green / 255.0, blue / 255.0, 1.0 - transmission);
+            Color = new Color(red / 255.0f, green / 255.0f, blue / 255.0f, 1.0f - transmission);
             Metallic = metallic;
             Roughness = roughness;
             Sheen = sheen;
@@ -52,7 +52,7 @@ namespace Plato.Geometry.Graphics
             SpecularReflectance = specularReflectance;
             SpecularTint = specularTint;
             IndexOfRefraction = indexOfRefraction;
-            SubsurfaceColor = new Color(subsurfaceRed / 255.0, subsurfaceGreen / 255.0, subsurfaceBlue / 255.0, 1.0);
+            SubsurfaceColor = new Color(subsurfaceRed / 255.0f, subsurfaceGreen / 255.0f, subsurfaceBlue / 255.0f, 1.0f);
             SubsurfaceScattering = subsurfaceScattering;
             Name = name;
         }
@@ -69,16 +69,16 @@ namespace Plato.Geometry.Graphics
     public struct SpecularSettings
     {
         public readonly Color Color;
-        public readonly double Power;
-        public SpecularSettings(Color color, double power)
+        public readonly float Power;
+        public SpecularSettings(Color color, float power)
         {
             Color = color;
             Power = power;
         }
 
-        public static Color DielectricSpecular = new Color(0.04, 0.04, 0.04, 1.0);
+        public static Color DielectricSpecular = new Color(0.04f, 0.04f, 0.04f, 1.0f);
 
-        public static SpecularSettings FromPBR(Color color, double metallic, double roughness)
+        public static SpecularSettings FromPBR(Color color, float metallic, float roughness)
             => new SpecularSettings(DielectricSpecular.Lerp(color, metallic), ((Number)1).Lerp(100, 1 - roughness));
     }
 }
